refactor: [FE] 메인페이지 소켓 코드 훅으로 분리
