+++ frontend/src/components/UserCard.tsx
@@ -6,14 +6,13 @@ import { motion } from 'framer-motion';
 import { opacityVariants } from '@utils/framerMotion';
 import { userState, userStreamState, userCamState } from '@atoms/user';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { networkServiceInstance as NetworkService } from '../services/socketService';
 import Card from '@components/Card';
 import CameraButton from '@components/CameraButton';
 import MicButton from '@components/MicButton';
 import useLocalStream from '@hooks/useLocalStream';
 import MainVideoCall from '@components/MainVideoCall';
 
-function UserCard() {
+function UserCard({ emitUpdateUserName }: { emitUpdateUserName: (name: string) => void }) {
     const [user, setUserState] = useRecoilState(userState);
     const userCam = useRecoilValue(userCamState);
     const currentUser = useRecoilValue(userState);
@@ -33,7 +32,7 @@ function UserCard() {
 
     const debounceOnChange = useCallback(
         debounce((name: string) => {
-            NetworkService.emit('update-user-name', name);
+            emitUpdateUserName(name);
         }, 500),
         [],
     );
